Moved MDAnalysisTests.core.groupbase to core.util
Made make_Universe available at MDAnalysisTests top namespace

Moved more tests out of test_atomgroup
