+++ testsuite/MDAnalysisTests/coordinates/test_netcdf.py
@@ -9,15 +9,12 @@ from numpy.testing import (assert_, assert_equal, assert_array_almost_equal,
                            assert_array_equal,
                            assert_almost_equal, assert_raises, dec)
 from unittest import TestCase
-from MDAnalysisTests import module_not_found
 
 from MDAnalysisTests.datafiles import (PRMncdf, NCDF, PFncdf_Top, PFncdf_Trj,
                                        GRO, TRR, XYZ_mini)
 from MDAnalysisTests.coordinates.test_trj import _TRJReaderTest
 from MDAnalysisTests.coordinates.reference import (RefVGV, RefTZ2)
-from MDAnalysisTests import tempdir
-from MDAnalysisTests.util import block_import
-from MDAnalysisTests.core.groupbase import make_Universe
+from MDAnalysisTests import module_not_found, tempdir, block_import, make_Universe
