+++ testsuite/MDAnalysisTests/core/test_topology.py
@@ -15,7 +15,7 @@ from numpy.testing import (
 )
 import numpy as np
 
-from MDAnalysisTests.core.groupbase import make_Universe
+from MDAnalysisTests import make_Universe
 
 from MDAnalysis.core.topology import (
     Topology,
