+++ testsuite/MDAnalysisTests/test_atomgroup.py
@@ -43,8 +43,7 @@ from unittest import skip
 import os
 import itertools
 
-from MDAnalysisTests import parser_not_found, tempdir
-from MDAnalysisTests.core.groupbase import make_Universe
+from MDAnalysisTests import parser_not_found, tempdir, make_Universe
 
 
 class TestUniverseSetTopology(object):
@@ -167,316 +166,6 @@ class TestUniverseSetTopology(object):
         assert_equal('improperDict' in self.u._cache, False)
 
 
-class TestResidue(object):
-    @dec.skipif(parser_not_found('DCD'),
-                'DCD parser not available. Are you using python 3?')
-    def setUp(self):
-        self.universe = MDAnalysis.Universe(PSF, DCD)
-        self.res = self.universe.residues[100]
-
-    def test_type(self):
-        assert_(isinstance(self.res, mda.core.groups.Residue))
-        assert_equal(self.res.resname, "ILE")
-        assert_equal(self.res.resid, 101)
-
-    # INVALID: residues do not behave like AtomGroups anymore,
-    # so cannot index in this way; should use `res.atoms[2]`
-    @skip
-    def test_index(self):
-        atom = self.res.atoms[2]
-        assert_equal(type(atom), mda.core.groups.Atom)
-        assert_equal(atom.name, "CA")
-        assert_equal(atom.index, 1522)
-        assert_equal(atom.resid, 101)
-
-    @skip
-    def test_slicing(self):
-        atoms = self.res[2:10:2]
-        assert_equal(len(atoms), 4)
-        assert_(isinstance(atoms, mda.core.groups.AtomGroup))
-
-    # INVALID: residues do not behave like AtomGroups anymore,
-    # so cannot slice in this way
-    @skip
-    def test_advanced_slicing(self):
-        atoms = self.res[[0, 2, -2, -1]]
-        assert_equal(len(atoms), 4)
-        assert_equal(type(atoms), MDAnalysis.core.groups.AtomGroup)
-        assert_equal(atoms.names, ["N", "CA", "C", "O"])
-
-    # VALID
-    def test_atom_order(self):
-        assert_equal(self.res.atoms.indices,
-                     sorted(self.res.atoms.indices))
-
-
-class TestResidueGroup(object):
-    @dec.skipif(parser_not_found('DCD'),
-                'DCD parser not available. Are you using python 3?')
-    # VALID
-    def setUp(self):
-        """Set up the standard AdK system in implicit solvent."""
-        self.universe = MDAnalysis.Universe(PSF, DCD)
-        self.rg = self.universe.atoms.residues
-
-    # VALID
-    def test_newResidueGroup(self):
-        """test that slicing a ResidueGroup returns a new ResidueGroup
-        (Issue 135)"""
-        rg = self.universe.atoms.residues
-        newrg = rg[10:20:2]
-        assert_equal(type(newrg), type(rg),
-                     "Failed to make a new ResidueGroup: type mismatch")
-        assert_equal(len(newrg), len(rg[10:20:2]))
-
-    # VALID
-    def test_n_atoms(self):
-        assert_equal(self.rg.n_atoms, 3341)
-
-    # VALID
-    def test_n_residues(self):
-        assert_equal(self.rg.n_residues, 214)
-
-    # VALID
-    def test_resids_dim(self):
-        assert_equal(len(self.rg.resids), len(self.rg))
-
-    # INVALID: this topology has no resnums, so no resnums property
-    @skip
-    def test_resnums_dim(self):
-        assert_equal(len(self.rg.resnums), len(self.rg))
-
-    # VALID
-    def test_segids_dim(self):
-        assert_equal(len(self.rg.segids), len(self.rg))
-
-    # VALID
-    def test_len(self):
-        """testing that len(residuegroup) == residuegroup.n_residues"""
-        assert_equal(len(self.rg), self.rg.n_residues,
-                     "len and n_residues disagree")
-
-    # INVALID: set resids with `ResidueGroup.resids` property; no `set_resids` method
-    @skip
-    def test_set_resids(self):
-        rg = self.universe.select_atoms("bynum 12:42").residues
-        resid = 999
-        rg.set_resids(resid)
-        # check individual atoms
-        assert_equal([a.resid for a in rg.atoms],
-                     resid * np.ones(rg.n_atoms),
-                     err_msg="failed to set_resid atoms 12:42 to same resid")
-        # check residues
-        assert_equal(rg.resids, resid * np.ones(rg.n_residues),
-                     err_msg="failed to set_resid of residues belonging to "
-                     "atoms 12:42 to same resid")
-
-    # INVALID: set resids with `ResidueGroup.resids` property; no `set_resids` method
-    @skip
-    def test_set_resids(self):
-        """test_set_resid: set ResidueGroup resids on a per-residue basis"""
-        rg = self.universe.select_atoms("resid 10:18").residues
-        resids = np.array(rg.resids) + 1000
-        rg.set_resids(resids)
-        # check individual atoms
-        for r, resid in zip(rg, resids):
-            assert_equal([a.resid for a in r.atoms],
-                         resid * np.ones(r.n_atoms),
-                         err_msg="failed to set_resid residues 10:18 to same "
-                         "resid in residue {0}\n"
-                         "(resids = {1}\nresidues = {2})".format(r, resids, rg))
-        # check residues
-        # NOTE: need to create a new selection because underlying Residue
-        #       objects are not changed; only Atoms are changed, and Residues
-        #       are rebuilt from Atoms.
-        rgnew = self.universe.select_atoms("resid 1010:1018").residues
-        assert_equal(rgnew.resids, np.unique(resids),
-                     err_msg="failed to set_resid of residues belonging to "
-                     "residues 10:18 to new resids")
-
-    # INVALID: set resids with `ResidueGroup.resids` property; no `set_resids` method
-    @skip
-    def test_set_resids_updates_self(self):
-        rg = self.universe.select_atoms("resid 10:18").residues
-        resids = np.array(rg.resids) + 1000
-        rg.set_resids(resids)
-        assert_equal(rg.resids, np.unique(resids),
-                     err_msg="old selection was not changed in place "
-                     "after set_resid")
-
-    # INVALID: no resnums in this topology, so no resnums property
-    @skip
-    def test_set_resnum_single(self):
-        rg = self.universe.residues[:3]
-        new = 22
-        rg.set_resnums(new)
-
-        assert_equal(all(rg.resnums == new), True)
-        for r in rg:
-            assert_equal(r.resnum, new)
-
-    # INVALID: no resnums in this topology, so no resnums property
-    @skip
-    def test_set_resnum_many(self):
-        rg = self.universe.residues[:3]
-        new = [22, 23, 24]
-        rg.set_resnums(new)
-
-        assert_equal(all(rg.resnums == new), True)
-        for r, v in zip(rg, new):
-            assert_equal(r.resnum, v)
-
-    # INVALID: no resnums in this topology, so no resnums property
-    @skip
-    def test_set_resnum_ValueError(self):
-        rg = self.universe.residues[:3]
-        new = [22, 23, 24, 25]
-
-        assert_raises(ValueError, rg.set_resnums, new)
-
-    # INVALID: no `set_resnames` method; use `resnames` property directly
-    @skip
-    def test_set_resname_single(self):
-        rg = self.universe.residues[:3]
-        new = 'newname'
-
-        rg.set_resnames(new)
-        assert_equal(all(rg.resnames == new), True)
-        for r in rg:
-            assert_equal(r.name, new)
-
-    # INVALID: no `set_resnames` method; use `resnames` property directly
-    @skip
-    def test_set_resname_many(self):
-        rg = self.universe.residues[:3]
-        new = ['a', 'b', 'c']
-        rg.set_resnames(new)
-
-        assert_equal(all(rg.resnames == new), True)
-        for r, v in zip(rg, new):
-            assert_equal(r.name, v)
-
-    # INVALID: no `set_resnames` method; use `resnames` property directly
-    @skip
-    def test_set_resname_ValueError(self):
-        rg = self.universe.residues[:3]
-        new = ['a', 'b', 'c', 'd']
-
-        assert_raises(ValueError, rg.set_resnames, new)
-
-    # INVALID: no `set_resids` method; also, residues are not mergeable
-    # by setting resids; resids are not necessarily unique; atoms must
-    # have their resindex set to change residue membership
-    @skip
-    def test_merge_residues(self):
-        rg = self.universe.select_atoms("resid 12:14").residues
-        nres_old = self.universe.atoms.n_residues
-        natoms_old = rg.n_atoms
-        rg.set_resids(12)  # merge all into one with resid 12
-        nres_new = self.universe.atoms.n_residues
-        r_merged = self.universe.select_atoms("resid 12:14").residues
-        natoms_new = self.universe.select_atoms("resid 12").n_atoms
-        assert_equal(len(r_merged), 1, err_msg="set_resid failed to merge "
-                     "residues: merged = {0}".format(r_merged))
-        assert_equal(nres_new, nres_old - 2,
-                     err_msg="set_resid failed to merge residues: "
-                     "merged = {0}".format(r_merged))
-        assert_equal(natoms_new, natoms_old, err_msg="set_resid lost atoms "
-                     "on merge".format(r_merged))
-
-        assert_equal(self.universe.residues.n_residues,
-                     self.universe.atoms.n_residues,
-                     err_msg="Universe.residues and Universe.atoms.n_residues "
-                     "do not agree after residue "
-                             "merge.")
-
-    # INVALID: no `set_masses` method; use `masses` property directly
-    @skip
-    def test_set_masses(self):
-        rg = self.universe.select_atoms("bynum 12:42 and name H*").residues
-        mass = 2.0
-        rg.set_masses(mass)
-        # check individual atoms
-        assert_equal([a.mass for a in rg.atoms],
-                     mass * np.ones(rg.n_atoms),
-                     err_msg="failed to set_mass H* atoms in resid 12:42 to {0}".format(mass))
-
-    # VALID
-    def test_atom_order(self):
-        assert_equal(self.universe.residues.atoms.indices,
-                     sorted(self.universe.residues.atoms.indices))
-
-
-class TestSegment(object):
-    @dec.skipif(parser_not_found('DCD'),
-                'DCD parser not available. Are you using python 3?')
-
-    # INVALID: no `set_segids` method for ResidueGroup; c
-    @skip
-    def setUp(self):
-        self.universe = MDAnalysis.Universe(PSF, DCD)
-        self.universe.residues[:100].set_segids("A")  # make up some segments
-        self.universe.residues[100:150].set_segids("B")
-        self.universe.residues[150:].set_segids("C")
-        self.sB = self.universe.segments[1]
-
-    # VALID but temporary
-    @dec.skipif(parser_not_found('DCD'),
-                'DCD parser not available. Are you using python 3?')
-    def setUp(self):
-        self.universe = MDAnalysis.Universe(PSF, DCD)
-
-    # INVALID: `Segment` from a particular Universe will not be the same class
-    # in core; each Universe generates its own set of classes based on topology
-    @skip
-    def test_type(self):
-        assert_equal(type(self.sB), MDAnalysis.core.groups.Segment)
-        assert_equal(self.sB.name, "B")
-
-    # INVALID: `Segment` doesn't behave as a `ResidueGroup`;
-    # no index behavior
-    @skip
-    def test_index(self):
-        s = self.sB
-        res = s[5]
-        assert_equal(type(res), MDAnalysis.core.groups.Residue)
-
-    # INVALID: `Segment` doesn't behave as a `ResidueGroup`;
-    # no slicing behavior
-    @skip
-    def test_slicing(self):
-        res = self.sB[5:10]
-        assert_equal(len(res), 5)
-        assert_equal(type(res), MDAnalysis.core.groups.ResidueGroup)
-
-    # INVALID: `Segment` doesn't behave as a `ResidueGroup`;
-    # no slicing behavior
-    @skip
-    def test_advanced_slicing(self):
-        res = self.sB[[3, 7, 2, 4]]
-        assert_equal(len(res), 4)
-        assert_equal(type(res), MDAnalysis.core.groups.ResidueGroup)
-
-    # INVALID: no `name` or `id` attribute for `Segment`; only `segid`
-    @skip
-    def test_id(self):
-        assert_equal(self.sB.name, self.sB.id)
-
-    # INVALID: no `name` or `id` attribute for `Segment`; only `segid`
-    @skip
-    def test_set_id(self):
-        # Test setting the name via the id attribute
-        new = 'something'
-        self.sB.id = new
-        for val in [self.sB.id, self.sB.name]:
-            assert_equal(val, new)
-
-    # VALID
-    def test_atom_order(self):
-        assert_equal(self.universe.segments[0].atoms.indices,
-                     sorted(self.universe.segments[0].atoms.indices))
-
-
 class TestSegmentGroup(object):
     # VALID
     @dec.skipif(parser_not_found('DCD'),
@@ -576,33 +265,6 @@ class TestSegmentGroup(object):
     def test_atom_order(self):
         assert_equal(self.universe.segments.atoms.indices,
                      sorted(self.universe.segments.atoms.indices))
-
-
-class TestAtomGroupTimestep(object):
-    """Tests the AtomGroup.ts attribute (partial timestep)"""
-
-    @dec.skipif(parser_not_found('TRZ'),
-                'TRZ parser not available. Are you using python 3?')
-    def setUp(self):
-        self.universe = MDAnalysis.Universe(TRZ_psf, TRZ)
-        self.prec = 6
-
-    def tearDown(self):
-        del self.universe
-        del self.prec
-
-    def test_partial_timestep(self):
-        ag = self.universe.select_atoms('name Ca')
-        idx = ag.indices
-
-        assert_equal(len(ag.ts._pos), len(ag))
-
-        for ts in self.universe.trajectory[0:20:5]:
-            assert_array_almost_equal(ts.positions[idx], ag.ts.positions, self.prec,
-                                      err_msg="Partial timestep coordinates wrong")
-            assert_array_almost_equal(ts.velocities[idx], ag.ts.velocities, self.prec,
-                                      err_msg="Partial timestep coordinates wrong")
-
 
 
 class _WriteAtoms(object):
