+++ testsuite/MDAnalysisTests/topology/test_guessers.py
@@ -31,7 +31,7 @@ import numpy as np
 from MDAnalysis.topology import guessers
 from MDAnalysis.core.topologyattrs import Angles
 
-from MDAnalysisTests.core.groupbase import make_Universe
+from MDAnalysisTests import make_Universe
 from MDAnalysisTests.core.test_fragments import make_starshape
 
 class TestGuessMasses(object):
